Move GLES' SoftwareTransform to Common
