+++ GPU/GPU.vcxproj.filters
@@ -171,6 +171,9 @@
     <ClInclude Include="Common\FramebufferCommon.h">
       <Filter>Common</Filter>
     </ClInclude>
+    <ClInclude Include="Common\SoftwareTransformCommon.h">
+      <Filter>Common</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Math3D.cpp">
@@ -207,9 +210,6 @@
       <Filter>GLES</Filter>
     </ClCompile>
     <ClCompile Include="GLES\TransformPipeline.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\SoftwareTransform.cpp">
       <Filter>GLES</Filter>
     </ClCompile>
     <ClCompile Include="GLES\VertexShaderGenerator.cpp">
@@ -321,6 +321,9 @@
       <Filter>Common</Filter>
     </ClCompile>
     <ClCompile Include="Common\VertexDecoderArm.cpp">
+      <Filter>Common</Filter>
+    </ClCompile>
+    <ClCompile Include="Common\SoftwareTransformCommon.cpp">
       <Filter>Common</Filter>
     </ClCompile>
   </ItemGroup>
